+++ lib/rich_form_field.dart
@@ -1,7 +1,10 @@
-library rich_form_field;
+library;
 
 import 'package:flutter/material.dart';
+import 'package:html/dom.dart' as html_dom;
+import 'package:html/parser.dart' as html_parser;
 
+part 'src/rich_text_codec.dart';
 part 'src/rich_text_editor_strings.dart';
 part 'src/html_rich_text_form_field.dart';
 part 'src/html_rich_text_controller.dart';